Fix memcpy types

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: pschwarz <pschwarz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/11 14:45:59 by pschwarz          #+#    #+#             */
-/*   Updated: 2022/10/19 16:53:12 by pschwarz         ###   ########.fr       */
+/*   Updated: 2022/10/19 17:27:29 by pschwarz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int				ft_isprint(int c);
 unsigned int	ft_strlen(const char *s);
 void			*ft_memset(void *b, int c, int len);
 void			*ft_bzero(void *s, int n);
-void			*ft_memcpy(void *dst, const void *src, int n);
+void			*ft_memcpy(void *dst, const void *src, size_t n);
 void			*ft_memmove(void *dst, const void *src, int len);
 unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int dstsize);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
